refactor(fe): 글 부분 분리
