+++ frontend/src/components/Post/PostHead.tsx
@@ -1,20 +1,9 @@
 /* 포스트 상단 */
 import styled from '@emotion/styled';
-import { evaluateSync } from '@mdx-js/mdx';
 import { postType } from 'api/apiTypes';
 import mdxComponents from 'editor/mdxComponent';
-import * as runtime from 'react/jsx-runtime';
-import React, { createElement } from 'react';
 import colors from 'styles/color';
 import { dateToStr } from 'utils/utils';
-
-const generate = (body: string) => {
-    const mdx = evaluateSync(body, {
-        ...(runtime as any),
-    }).default;
-
-    return createElement(mdx);
-};
 
 interface PostHeadPropsType {
     postInfo: postType;
@@ -34,9 +23,6 @@ const PostHead = ({ postInfo }: PostHeadPropsType) => {
                     <p>{postInfo.tagId}</p>
                 </TagArea>
             </PostInfo>
-            <ContentArea>
-                {generate(mdxComponents + postInfo.content)}
-            </ContentArea>
         </>
     );
 };
@@ -67,15 +53,6 @@ const DateArea = styled.p`
     p {
         margin-right: 0.5rem;
     }
-`;
-
-const ContentArea = styled.div`
-    width: 60vw;
-    margin-left: calc(20vw - 150px);
-    margin-top: 3rem;
-    position: relative;
-    padding: 3rem;
-    border: 1px solid ${colors.boxGrayLine};
 `;
 
 export default PostHead;
